Fix integer out of range in mt_quick_append_events and mt_get_next_hi SQL functions
The mt_quick_append_events function declared its return type and internal
variables as int/int[] while the underlying sequence (mt_events_sequence)
and seq_id column use bigint. This causes integer overflow when sequence
IDs exceed ~2.1 billion, which was observed in production.

Changed in mt_quick_append_events:
- RETURNS int[] → bigint[]
- event_version int → bigint
- seq int → bigint
- return_value int[] → bigint[]

Also fixed mt_get_next_hi which declared RETURNS integer despite its
hi_value column being bigint and its internal variables being bigint.

C# consumers already read these values as long/long[], so no client-side
changes are needed. A test is added that advances the event sequence past
int.MaxValue and verifies quick append still works correctly.

+++ src/Marten/Schema/SQL/mt_get_next_hi.sql
@@ -1,5 +1,5 @@
 ﻿CREATE
-OR REPLACE FUNCTION {databaseSchema}.mt_get_next_hi(entity varchar) RETURNS integer AS
+OR REPLACE FUNCTION {databaseSchema}.mt_get_next_hi(entity varchar) RETURNS bigint AS
 $$
 DECLARE
 current_value bigint;
